HW-1 Make Movie class nested in MovieResponse class
