--- class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/model/Movie.kt
@@ -1,11 +1,0 @@
-package com.example.googryandroidarchitecturestudy.model
-
-data class Movie (
-    val title: String,
-    val image: String,
-    val subtitle: String,
-    val pubDate: String,
-    val director: String,
-    val actor: String,
-    val userRating: String
-)
