--- class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/model/MovieResponse.kt
@@ -1,9 +1,0 @@
-package com.example.googryandroidarchitecturestudy.model
-
-data class MovieResponse(
-    val display: Int,
-    val items: List<Movie>,
-    val lastBuildDate: String,
-    val start: Int,
-    val total: Int
-)
