+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/network/MovieApiService.kt
@@ -1,7 +1,6 @@
 package com.example.googryandroidarchitecturestudy.network
 
-import com.example.googryandroidarchitecturestudy.model.MovieResponse
-import retrofit2.Call
+import com.example.googryandroidarchitecturestudy.data.model.MovieResponse
 import retrofit2.http.GET
 import retrofit2.http.Query
