+++ class14/juan-rybczinski/GoogryAndroidArchitectureStudy/app/src/main/java/com/example/googryandroidarchitecturestudy/ui/recycler/MovieViewHolder.kt
@@ -1,13 +1,12 @@
 package com.example.googryandroidarchitecturestudy.ui.recycler
 
-import android.text.Html
 import android.view.View
 import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.example.googryandroidarchitecturestudy.R
-import com.example.googryandroidarchitecturestudy.model.Movie
+import com.example.googryandroidarchitecturestudy.data.model.MovieResponse.Movie
 
 class MovieViewHolder(view: View) : RecyclerView.ViewHolder(view) {
